+++ WebUI/src/components/SettingsApp.vue
@@ -1,0 +1,8 @@
+<template>
+  <div>
+    App Settings
+  </div>
+</template>
+<script setup lang="ts">
+</script>
+
