+++ WebUI/src/components/SideModalHistory.vue
